pyproject.toml
[project]
name = "adk-web-multi-agent"
version = "0.1.0"
description = "Multi Agent System for Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "litellm>=1.78.7",
    "toolbox-core>=0.5.2",
]
